Add support for parsing /proc/self/mountinfo
Follows the same code structure as existing support. A MountInfoIter can
be created from either /proc/self/mountinfo or /proc/<pid>/mountinfo.

There are some basic tests but not quite as thorough and nice as the
mounts tests.

Add errors to error::LineError as needed for the different fields present
in mountinfo.

Signed-off-by: Andy Grover <agrover@redhat.com>